Update SoccerLeague_Grimes_Chirokav.sql
This should be everything in terms of tables and data. I switched the values of stats for clubs and players to a default value of zero. I'm thinking that every club and player should start off with no values in their stats so that only stored procedures can provide numbers to fill such stat data.